Se arregla e instalan dependencias.
